requirements.txt
 pandas
 plotly_express
 streamlits